% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{dpp_pc1}
\alias{dpp_pc1}
\title{Extract first principal component}
\usage{
dpp_pc1(data, ..., name = "pc1", center = T, scale. = F, keep = T)
}
\arguments{
\item{data}{A tibble (recommended) or a data frame.}

\item{...}{<\link[dplyr]{dplyr_tidy_select}>. Columns to calculate the first PC.}

\item{name}{A string of the new variable name. Default is "pc1".}

\item{center}{Pass to \link[stats]{prcomp}.}

\item{scale.}{Pass to \link[stats]{prcomp}.}

\item{keep}{Whether to keep columns, based on which the PC is calculated, in the resulted data.}
}
\value{
A data frame with a new column named \code{name}, which value is the first PC
of \code{...} columns.
}
\description{
Besides taking arithmetic mean value or weighted average, another useful way
to reduce dimensionality or construct an index out of several variables is
\strong{Principal Component} or PC. The first PC contains largest information and
variation of original data. \code{dpp_pc1()} provides a one-line solution to extract
the first PC and store it in a new column.
}
\examples{
dpp_pc1(iris, matches("Petal"), name = "avg_petal", keep = F)

}
