% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{sum_na}
\alias{sum_na}
\title{Tell you how many NAs are in your data}
\usage{
sum_na(data, ..., by = NULL)
}
\arguments{
\item{data}{A tibble (recommended), or a data frame.}

\item{...}{<\link[dplyr]{dplyr_tidy_select}>. Columns to summarize.}

\item{by}{(Optional) Columns to group by for this summarize operation.}
}
\value{
A data frame with columns in \code{...}. Each row corresponds to a group in
\code{by} variables. The value is the number of NAs in each group divided by total
observations in that group.
}
\description{
\code{sum_na()} explicitly shows the proportion of NA for each variable you select
in your data (in \%). You can also summarize this information by group variables.
If the data is wide (having a large number of columns), I strongly recommend
to use a \code{glimpse()} after using \code{sum_na()}.
}
\examples{
data <- tibble(x1 = c(1, 2, 2), x2 = c(1, NA, 23), y = c("a", NA, NA))
sum_na(data)    # All variables
sum_na(data, x1, x2, y)
sum_na(data, matches("x"))
sum_na(data, -y)
sum_na(data, x2, y, by = x1)
}
