% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{tab_validnum}
\alias{tab_validnum}
\title{Unify valid number}
\usage{
tab_validnum(x, valid = 4)
}
\arguments{
\item{x}{A numeric vector.}

\item{valid}{Valid numbers.}
}
\description{
People want their results in summary statistics or regression tables to have
uniform valid numbers. However, \code{round()} and \code{signif()} in \strong{baseR} will produce
different results when \code{x} cross 1. \code{tab_validnum} handles this by adding one
to the digits in \code{signif()} if \code{x} is greater than 1.
}
\examples{
tab_validnum(0.0423, 3); tab_validnum(1.0423, 3)
tab_validnum(1.4234, 3); tab_validnum(22.4234, 3)
}
